Set git config push.default to current
Git is about to change the default behavior in 2.0. The current default
behavior is 'matching', which pushes all branches that have the same
name as the remote. Git 2.0 is changing this to 'simple' which only
updates the current branch if it's name matches the remote.

'current' pushes the current branch even if the two names don't match.
This is my desired behavior since I often will name my branches
something different on the server, or I have 2+ local branches tracking
one remote branch.