Set Conjure log to behave like vim-fireplace
Currently Conjure logs put themselves into the NE corner and take up
approximently 2/5 of the width and 1/3 height of the current panel. This
can obscure code when I'm using a vertical monitor, especially one that
is 1080p, as well as truncate output that I would like to read.

This changes the log to put itself on the bottom of the panel at 1/5
height, mimicking vim-fireplace. The panel relocates to the top when the
cursor goes near it, preventing it from obscuring more code and makeing
it always readable for longer output strings.